+++ codegen/shim/src/lib.rs
@@ -116,12 +116,15 @@ use arcane as _;
 // Only for generating documentation.
 #[cfg(feature = "doc")]
 use arcane_core as _;
-#[cfg(doc)]
+#[cfg(all(doc, feature = "doc", feature = "es"))]
 use arcane_core::es::event;
 
+#[cfg(feature = "es")]
 use arcane_codegen_impl as codegen;
+#[cfg(feature = "es")]
 use proc_macro::TokenStream;
 
+#[cfg(feature = "es")]
 /// Macro for deriving the [`Event`] trait on structs and enums.
 ///
 /// # Enums
@@ -138,7 +141,7 @@ use proc_macro::TokenStream;
 /// Also, provides a blanket [`event::Sourced`] implementation for every state,
 /// which can be sourced from all the enum variants.
 ///
-/// > **WARNING:** Currently may not work with complex generics using where
+/// > **WARNING**: Currently may not work with complex generics using where
 /// >              clause because of `const` evaluation limitations. Should be
 /// >              lifted once [rust-lang/rust#57775] is resolved.
 ///
@@ -166,7 +169,7 @@ use proc_macro::TokenStream;
 /// Use this on a particular enum variant to completely ignore it in code
 /// generation.
 ///
-/// > **WARNING:** Calling [`Event::name()`] or [`event::Revisable::revision()`]
+/// > **WARNING**: Calling [`Event::name()`] or [`event::Revisable::revision()`]
 /// >              on ignored variants will result in [`unreachable!`] panic.
 ///
 /// ## Example